@@ -5,34 +5,39 @@
     "dev": "next dev",
     "build": "next build",
     "start": "next start",
-    "lint": "eslint src --ext js,ts,tsx"
+    "lint": "next lint"
   },
   "license": "Unlicense",
   "devDependencies": {
-    "@types/md5": "^2.3.0",
-    "@types/mongodb": "^3.6.8",
-    "@types/node": "^14.14.31",
+    "@next/eslint-plugin-next": "^12.0.10",
+    "@types/md5": "^2.3.1",
+    "@types/mongodb": "^4.0.7",
+    "@types/node": "^17.0.15",
     "@types/nprogress": "^0.2.0",
-    "@typescript-eslint/eslint-plugin": "^4.16.1",
-    "@typescript-eslint/parser": "^4.16.1",
-    "eslint": "^7.20.0",
-    "eslint-config-prettier": "^8.0.0",
-    "eslint-plugin-prettier": "^3.3.1",
-    "eslint-plugin-react": "^7.22.0",
+    "@typescript-eslint/eslint-plugin": "^5.10.2",
+    "@typescript-eslint/parser": "^5.10.2",
+    "eslint": "^8.8.0",
+    "eslint-config-prettier": "^8.3.0",
+    "eslint-plugin-prettier": "^4.0.0",
+    "eslint-plugin-react": "^7.28.0",
     "eslint-plugin-simple-import-sort": "^7.0.0",
-    "prettier": "^2.2.1",
-    "sass": "^1.29.0",
-    "typescript": "^4.2.2"
+    "prettier": "^2.5.1",
+    "sass": "^1.49.7",
+    "typescript": "^4.5.5",
+    "yarn-upgrade-all": "^0.6.1"
   },
   "dependencies": {
-    "@material-ui/core": "^4.11.1",
-    "@material-ui/icons": "^4.9.1",
+    "@emotion/react": "^11.7.1",
+    "@emotion/styled": "^11.6.0",
+    "@mui/icons-material": "^5.3.1",
+    "@mui/material": "^5.4.0",
+    "add": "^2.0.6",
     "md5": "^2.3.0",
-    "mongodb": "^3.6.4",
-    "next": "^10.0.7",
-    "next-pwa": "^5.0.6",
+    "mongodb": "^4.3.1",
+    "next": "^12.0.10",
+    "next-pwa": "^5.4.4",
     "nprogress": "^0.2.0",
-    "react": "^17.0.1",
-    "react-dom": "^17.0.1"
+    "react": "^17.0.2",
+    "react-dom": "^17.0.2"
   }
 }
